Move regex files to validate package
